Pick Up Logic for Active/Passive Items
The input of pick up (Key F on the keyboard) and following operations are added to the BP_ThirdPersonCharacter
There are base blueprints for both Active and Passive items, we can create a new item by creating child blueprint on that base.
For more detailed logic:
Only one active item can be equipped, picking up a new active item will have the equipped one swapped, but it will get spawned again in front of the player character.
Passive items are stored in an array called EquippedPassiveItems, it does not have a size limitation. If the player pick up a new passive item, it will be added to the array with a new index. If the picked up item already exist in the array, it will add one level to that passive item (The level here can be used to upgrade the effect of the passive item)

Next steps:
1. Realize a function to USE active items.
2. Basic widget (HUD) to show the equipped item