Update from main& Add a copy of chest that can add weapon to inventory
The inventory is just inside the ThirdPersonCharacter blueprint now, it just works now, maybe they can be moved to a new independent blueprint.
The variable "EquippedWeapons" is an array of Weapon Data, it stores all equipped weapons of the player in an array. I copied a chest to test if that event "Add WeaponToInventory" works. Sandeep made his logic for picking up weapons in ThirdPersonCharacter blueprint, we may need further discussion.