Item System - Update AddToInventory custom events
The inventory is just inside the ThirdPersonCharacter blueprint now, it just works, but maybe they can be moved to a new independent blueprint.
The variable "EquippedWeapons" is an array of Weapon Data, it stores all equipped weapons of the player in an array. I copied a chest to test if the event "Add WeaponToInventory" works. Sandeep also added his logic for picking up weapons in the ThirdPersonCharacter blueprint, we may need further discussion.